back/requirements.txt
llama-cpp-python 
flask 
flask-cors
huggingface_hub
python-dotenv
gunicorn